Fix docker startup script (#67)
yarn start and bash were complaining about foreman not being found.
(Guessing it was the && operator in the docker CMD). Moved the foreman
install into the Dockerfile and updated the docker CMD in startup
scripts.

I wasn't able to troubleshoot from my machine. (I'm out of disk space!)
Paired with @karthik-sista and @melbourne.